Moves CloseIcon onto Input
